### 2025-08-04 #8723 @editorialbot
[PRE REVIEW]: TarGene: A Nextflow pipeline for the estimation of genetic effects on human traits via semi-parametric methods.
**Submitting author:** @olivierlabayle (<a href="http://orcid.org/0000-0002-3708-3706">Olivier Labayle</a>)
**Repository:** https://github.com/TARGENE/targene-pipeline
**Branch with paper.md** (empty if default branch):
**Version:** v0.12.0
