+++ .github/workflows/CI.yml
@@ -40,6 +40,16 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
       - uses: nf-core/setup-nextflow@v1.2.0
+      - name: Maximize build space
+        uses: easimon/maximize-build-space@master
+        if: contains(matrix.testrun, 'simulation')
+        with:
+          root-reserve-mb: 512
+          swap-size-mb: 1024
+          remove-dotnet: 'true'
+          remove-android: 'true'
+          remove-haskell: 'true'
+          remove-codeql: 'true'
       - name: Instantiate test environment
         run: julia --project=test --startup-file=no -e'using Pkg;Pkg.instantiate()'
       - name: Run tests
