### 2025-11-01 2dc49d9
fix placement of cleanup action
