### 2025-11-01 d0c6477
add subset mode; prevalence update; new CLI
