### 2025-11-03 5041024
attmepting to update deps; bugfix in targenecore for reading files
