### 2025-11-06 #8723 @abhishektiwari
Also @olivierlabayle there seems to be a [few missing DOIs](https://github.com/openjournals/joss-reviews/issues/8723#issuecomment-3151037426) for references in `paper.bib`. Can you please also review suggested DOIs and if needed update `paper.bib`?

```
❌ MISSING DOIs

- 10.1038/s41467-021-27930-0 may be a valid DOI for title: DeepNull models non-linear covariate effects to im...
- 10.1038/nbt.3820 may be a valid DOI for title: Nextflow enables reproducible computational workfl...
- 10.1038/s41588-018-0248-z may be a valid DOI for title: An atlas of genetic associations in UK Biobank
- 10.1038/s41588-018-0144-6 may be a valid DOI for title: Mixed-model association for biobank-scale datasets
- 10.1038/s41588-018-0184-y may be a valid DOI for title: Efficiently controlling for case-control imbalance...
- 10.1056/nejmsr1809937 may be a valid DOI for title: The “All of Us” research program

```
