+++ paper/paper.bib
@@ -1,5 +1,6 @@
 @article{mbatchou2021computationally,
   title={Computationally efficient whole-genome regression for quantitative and binary traits},
+  doi={0.1038/s41588-021-00870-7},
   author={Mbatchou, Joelle and Barnard, Leland and Backman, Joshua and Marcketta, Anthony and Kosmicki, Jack A and Ziyatdinov, Andrey and Benner, Christian and O’Dushlaine, Colm and Barber, Mathew and Boutkov, Boris and others},
   journal={Nature genetics},
   volume={53},
@@ -11,6 +12,7 @@
 
 @article{mccaw2022deepnull,
   title={DeepNull models non-linear covariate effects to improve phenotypic prediction and association power},
+  doi={10.1038/s41467-021-27930-0},
   author={McCaw, Zachary R and Colthurst, Thomas and Yun, Taedong and Furlotte, Nicholas A and Carroll, Andrew and Alipanahi, Babak and McLean, Cory Y and Hormozdiari, Farhad},
   journal={Nature communications},
   volume={13},
@@ -22,6 +24,7 @@
 
 @article{sesia2021false,
   title={False discovery rate control in genome-wide association studies with population structure},
+  doi={10.1073/pnas.2105841118},
   author={Sesia, Matteo and Bates, Stephen and Cand{\`e}s, Emmanuel and Marchini, Jonathan and Sabatti, Chiara},
   journal={Proceedings of the National Academy of Sciences},
   volume={118},
@@ -33,6 +36,7 @@
 
 @article{di2017nextflow,
   title={Nextflow enables reproducible computational workflows},
+  doi={10.1038/nbt.3820},
   author={Di Tommaso, Paolo and Chatzou, Maria and Floden, Evan W and Barja, Pablo Prieto and Palumbo, Emilio and Notredame, Cedric},
   journal={Nature biotechnology},
   volume={35},
@@ -44,6 +48,7 @@
 
 @article{canela2018atlas,
   title={An atlas of genetic associations in UK Biobank},
+  doi={10.1038/s41588-018-0248-z},
   author={Canela-Xandri, Oriol and Rawlik, Konrad and Tenesa, Albert},
   journal={Nature genetics},
   volume={50},
@@ -55,6 +60,7 @@
 
 @inproceedings{chen2016xgboost,
   title={Xgboost: A scalable tree boosting system},
+  doi={10.1145/2939672.293978},
   author={Chen, Tianqi and Guestrin, Carlos},
   booktitle={Proceedings of the 22nd acm sigkdd international conference on knowledge discovery and data mining},
   pages={785--794},
@@ -63,6 +69,7 @@
 
 @book{van2018targeted,
   title={Targeted learning in data science},
+  doi={10.1007/978-3-319-65304-4},
   author={Van der Laan, Mark J and Rose, Sherri},
   year={2018},
   publisher={Springer}
@@ -70,6 +77,7 @@
 
 @article{purcell2007plink,
   title={PLINK: a tool set for whole-genome association and population-based linkage analyses},
+  doi={10.1086/519795},
   author={Purcell, Shaun and Neale, Benjamin and Todd-Brown, Kathe and Thomas, Lori and Ferreira, Manuel AR and Bender, David and Maller, Julian and Sklar, Pamela and De Bakker, Paul IW and Daly, Mark J and others},
   journal={The American journal of human genetics},
   volume={81},
@@ -81,6 +89,7 @@
 
 @article{yang2011gcta,
   title={GCTA: a tool for genome-wide complex trait analysis},
+  doi={10.1016/j.ajhg.2010.11.011},
[…]
