### 2025-11-17 5df848b
update dois and word count
