+++ paper/paper.bib
@@ -156,7 +156,7 @@
   publisher={Mass Medical Soc}
 }
 
-@article{10.1093/biostatistics/kxaf030,
+@article{labayle2025semi,
     author = {Labayle, Olivier and Roskams-Hieter, Breeshey and Slaughter, Joshua and Tetley-Campbell, Kelsey and van der Laan, Mark J and Ponting, Chris P and Beentjes, Sjoerd V and Khamseh, Ava},
     title = {Semiparametric efficient estimation of small genetic effects in large-scale population cohorts},
     journal = {Biostatistics},
