+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{mbatchou2021computationally,
   title={Computationally efficient whole-genome regression for quantitative and binary traits},
-  doi={0.1038/s41588-021-00870-7},
+  doi={10.1038/s41588-021-00870-7},
   author={Mbatchou, Joelle and Barnard, Leland and Backman, Joshua and Marcketta, Anthony and Kosmicki, Jack A and Ziyatdinov, Andrey and Benner, Christian and O’Dushlaine, Colm and Barber, Mathew and Boutkov, Boris and others},
   journal={Nature genetics},
   volume={53},
@@ -60,7 +60,7 @@
 
 @inproceedings{chen2016xgboost,
   title={Xgboost: A scalable tree boosting system},
-  doi={10.1145/2939672.293978},
+  doi={10.1145/2939672.2939785},
   author={Chen, Tianqi and Guestrin, Carlos},
   booktitle={Proceedings of the 22nd acm sigkdd international conference on knowledge discovery and data mining},
   pages={785--794},
