### 2025-11-17 dc5de4f
restrict word count even further
