### 2025-11-17 #8723 @olivierlabayle
@abhishektiwari Thank you for your help with this, the DOIs and word count are now up to date.
