### 2025-12-02 #8723 @kescobo
Hi @abhishektiwari - I'm not super familiar with the particular stats/human genetics side, but have lots of experience with Julia and nextflow. If you can find another reviewer that can speak more to the stats, I'm happy to review the implementation.
