+++ docs/src/examples/gwas.md
@@ -1,23 +1,27 @@
 # GWAS
 
-Perhaps the most common study design in population genetics is the Genome-Wide Association Study (GWAS). In TarGene, a GWAS can only be performed across [plink BED files](https://zzz.bwh.harvard.edu/plink/binary.shtml). A minimalist run configuration for a GWAS in TarGene looks like the following:
+Perhaps the most common study design in population genetics is the Genome-Wide Association Study (GWAS). In TarGene, a GWAS can only be performed across [plink BED files](https://zzz.bwh.harvard.edu/plink/binary.shtml). For our GWAS example, let's save the following snippet to a `my_gwas.conf` file:
 
 ```conf
 params {
-    ESTIMANDS_CONFIG = "gwas_config.yaml"
+    ESTIMANDS_CONFIG = "test/assets/gwas_config.yaml"
     ESTIMATORS_CONFIG = "wtmle--glm"
 
     // UK-Biobank specific parameters
-    BED_FILES = "unphased_bed/ukb_chr{1,2,3}.{bed,bim,fam}"
-    UKB_CONFIG = "ukbconfig_gwas.yaml"
-    TRAITS_DATASET = "dataset.csv"
-    UKB_WITHDRAWAL_LIST = "withdrawal_list.txt"
+    BED_FILES = "test/assets/unphased_bed/ukb_chr{1,2,3}.{bed,bim,fam}"
+    UKB_CONFIG = "test/assets/ukbconfig_gwas.yaml"
+    TRAITS_DATASET = "test/assets/dataset.csv"
+    UKB_WITHDRAWAL_LIST = "test/assets/withdrawal_list.txt"
 }
 ```
 
-Apart from the data related parameters, there are two main parameters here: `ESTIMANDS_CONFIG` and the `ESTIMATORS_CONFIG`. These parameters describe the estimands (questions of interest) and how to estimate them respectively. Since we are performing a GWAS, we are interested in the effect of all variants, across all phenotypes in the `UKB_CONFIG` file. 
+Let's now unpack what this configuration file is saying.
 
-The `ESTIMANDS_CONFIG` is a pretty succinct YAML file for a GWAS, it could contain only one line. Here we will be a little more fancy and add extra outcome predictors to improve the precision of inference
+- `BED_FILES`: This parameter points to the plink bed files containing genetic variants information. Notice the "{1,2,3}.{bed,bim,fam}" suffix of the parameter which is telling Nextflow to group chromosome files together.
+- `ESTIMATORS_CONFIG`: Defines the estimation strategy. We will be using Targeted Minimum-Loss Estimator with a simple Generalized Linear Model (glm) to learn the outcome models (``Q_Y``)and propensity scores (``G``). See [Specifying a Targeted Estimator](@ref) for more information on this.
+- `TRAITS_DATASET`: This is a CSV file containing both covariates and phenotypes of interest.It should contain an `eid` column to uniquely identify individuals and matching the plink individuals' `IID`.
+- `UKB_WITHDRAWAL_LIST`: List of individuals to be removed from the analysis.
+- `ESTIMANDS_CONFIG`: Since we are performing a GWAS, we are interested in the effect of all variants, across all phenotypes in the `UKB_CONFIG` file.  The `ESTIMANDS_CONFIG` is a pretty succinct YAML file for a GWAS, it could contain only one line. Here we will be a little more fancy and add extra outcome predictors to improve the precision of inference. The `test/assets/gwas_config.yaml` contains the following which is self explanatory.
 
 ```yaml
 type: gwas
@@ -27,7 +31,9 @@ outcome_extra_covariates:
   - "Cheese intake"
 ```
 
[…]
