+++ previews/PR219/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"all_workflows_parameters/#Index-of-Workflows-Parameters","page":"Index of Workflows Parameters","title":"Index of Workflows Parameters","text":"List of all workflows parameters. For more information on them please visit the dedicated user guides.","category":"section"},{"location":"all_workflows_parameters/#Required-Parameters","page":"Index of Workflows Parameters","title":"Required Parameters","text":"These must be provided, or the workflow will not run.\n\nESTIMANDS_CONFIG: YAML configuration file describing the effect sizes of interest.\nTRAITS_DATASET: Path to a traits dataset. If you are running this for a non-UKB cohort, your sample IDs must be specified in the first column of this CSV file, with the column name SAMPLE_ID.\nBED_FILES: Path to PLINK BED files.\nBGEN_FILES: Path to indexed imputed BGEN files (optional for a GWAS).","category":"section"},{"location":"all_workflows_parameters/#Main-Options","page":"Index of Workflows Parameters","title":"Main Options","text":"These are optional but important as they can have a significant impact on the workflow (speed, estimates, ...).\n\nESTIMATORS_CONFIG (default: wtmle-ose--tunedxgboost): Estimator name or Julia file containing the description of the Targeted Estimators to use. To be consistent it should match the argument provided to the previous TarGene run.\nBATCH_SIZE (default: 50): The set of estimands to be estimated is batched and the Targeted Learning processes will run in parallel across batches. This is the main driver of computational speed on High Performance Computing Platforms.\nCOHORT (default: UKB): Current default for this is UKB. If set to a value other than UKB, this will not run UKB-specific trait extraction.\nPOSITIVITY_CONSTRAINT (default: 0.01): When the list of estimands is generated or validated. Treatment variables' rarest configuration should have at least that frequency. For example if the treatment variables are two variants with minor allele A and T respectively. The rarest configuration will be (AA, TT) and should have a frequency of at least POSITIVITY_CONSTRAINT.\nNB_PCS (default: 6): The number of PCA components to extract.","category":"section"},{"location":"all_workflows_parameters/#UK-Biobank-Specific","page":"Index of Workflows Parameters","title":"UK Biobank Specific","text":"For UK Biobank analyses (COHORT=UKB).\n\nUKB_CONFIG (default: ${projectDir}/assets/ukbconfig.yaml): YAML configuration file describing which traits should be extracted and how the population should be subsetted.\nUKB_ENCODING_FILE: If the TRAITS_DATASET is encrypted, an encoding file must be provided.\nUKB_WITHDRAWAL_LIST: List of participants withdrawn from the study.\nQC_FILE: Genotyping quality control file from the UK-Biobank study.","category":"section"},{"location":"all_workflows_parameters/#Secondary-Options","page":"Index of Workflows Parameters","title":"Secondary Options","text":"These are of less importance.\n\nSVP (default: false): Whether Sieve Variance Plateau correction should be performed.\nFLASHPCA_EXCLUSION_REGIONS (default: assets/exclusion_regions_hg19.txt): A path to the flashpca special exclusion regions.\nMAF_THRESHOLD (default: 0.01): Only variants with that minor allele frequency are considered for PCA.\nLD_BLOCKS: A path to pre-identified linkage disequilibrium blocks to be removed from the BED files for PCA. It is good practice to specify LD_BLOCKS, as it will remove SNPs correlated with your variants-of-interest before running PCA.\nVERBOSITY (default: 0): Verbosity level of the the Workflow's processes.\nTL_SAVE_EVERY (default: BATCH_SIZE): During the estimation process, results are appended to the file in chunks to free m
[…]
