### 2026-01-13 9cfd4c0
update examples docs and container tags
