+++ docs/src/examples/phewas.md
@@ -1,8 +1,11 @@
 # PheWAS
 
-A phenome-wide association study (PheWAS) is a study design whereby the effect of a variant, or set of variants is estimated for a large number of traits or outcomes. In TarGene, running a PheWAS is always done implicitely because the outcomes are defined as all variables in the `TRAITS_DATASET` that do not play a particular role as per the `ESTIMANDS_FILE`.
+!!! note "Read First"
+  Make sure you have read the [Setup (Read First)](@ref) section.
 
-To run a PheWAS for 3 genetic variants we will use the `flat` configuration mode. Let's create an `ESTIMANDS_FILE` called "new_phewas_config.yaml" with the following content.
+A phenome-wide association study (PheWAS) is a study design whereby the effect of a variant, or set of variants is estimated for a large number of traits or outcomes. In TarGene, running a PheWAS is always done implicitely because the outcomes are defined as all variables in the `TRAITS_DATASET` that do not play a particular role in the `ESTIMANDS_FILE`. This file is a description of the target quantities you wish to estimate (as opposed to how you want to estimate them).
+
+To run a PheWAS for 3 genetic variants we will use the `flat` configuration mode. Let's create an `ESTIMANDS_FILE` called `new_phewas_config.yaml` with the following content.
 
 ```yaml
 type: flat
@@ -20,23 +23,21 @@ outcome_extra_covariates:
   - "Cheese intake"
 ```
 
-The `estimands` section is set to include Average Treatment Effects (`ATE`) which correspond to single variant effects. Because there are 3 variants in the `variants` section, 3 PheWAS will actually be run in paralllel. These will be run for all traits in the `UKB_CONFIG` that are not in the `outcome_extra_covariates` section.
+As a reminder, the `ESTIMANDS_FILE` describes the study design in TarGene. The `estimands` section is set to include Average Treatment Effects (`ATE`) which correspond to single variant effects (the counterpart to the beta coefficients in a linear model). Because there are 3 variants in the `variants` section, 3 PheWAS will actually be run in paralllel. These will be run for all traits in the `UKB_CONFIG` that are not in the `outcome_extra_covariates` section describing additional covariates used to fit the machine-learning models used by TarGene.
 
 !!! note "Note on variant IDs"
     The variant IDs must match the IDs of your BGEN files. Here they are identified by chr:pos:ref:alt but in your files it may be via the rsID. If this is the case, multi-allelic variants are currently not supported.
 
-The `nextflow.config` file for this study is:
+We can save the following snippet to fully describe our Nextflow run in `my_phewas.config`:
 
 ```conf
 params {
     ESTIMANDS_CONFIG = "new_phewas_config.yaml"
-
-    // UK-Biobank specific parameters
-    BED_FILES = "unphased_bed/ukb_chr{1,2,3}.{bed,bim,fam}"
-    BGEN_FILES = "unphased_bgen/ukb_chr{1,2,3}.{bgen,bgen.bgi,sample}"
-    UKB_CONFIG = "ukbconfig_small.yaml"
-    TRAITS_DATASET = "dataset.csv"
-    UKB_WITHDRAWAL_LIST = "withdrawal_list.txt"
+    BED_FILES = "test/assets/unphased_bed/ukb_chr{1,2,3}.{bed,bim,fam}"
+    BGEN_FILES = "test/assets/unphased_bgen/ukb_chr{1,2,3}.{bgen,bgen.bgi,sample}"
+    UKB_CONFIG = "test/assets/ukbconfig_small.yaml"
[…]
