+++ docs/src/examples/interactions.md
@@ -1,6 +1,9 @@
-# Interaction Study
+# Targeted Interaction Study
 
-In TarGene you can investigate the interacting effects between genetic variants and between genetic variants and the environment up to any order (GxG, GxE, GxGxG, ...). However, the computational complexity associated with interaction testing grows exponentially. Interaction studies are thus very often targeted, for instance towards a specific transcription factor `TF`. For this transcription factor, multiple variants may have been identified because they are believed to play different roles. In TarGene such a scenario can be encoded via the `group` mode of the `ESTIMANDS_CONFIG` file. An example is more informative than a long description. Let's save the following content to a "new_interaction_config.yaml" file.
+!!! note "Read First"
+  Make sure you have read the [Setup (Read First)](@ref) section.
+
+In TarGene you can investigate the interacting effects between genetic variants and between genetic variants and the environment up to any order (GxG, GxE, GxGxG, ...). However, the computational complexity associated with interaction testing grows exponentially. Interaction studies are thus very often targeted, for instance towards a specific transcription factor `TF`. For this transcription factor, multiple variants may have been identified because they are believed to play different roles. In TarGene such a scenario can be encoded via the `group` mode of the `ESTIMANDS_CONFIG` file. An example is more informative than a long description. Let's save the following content to a `new_interaction_config.yaml` file.
 
 ```yaml
 type: groups
@@ -24,36 +27,72 @@ outcome_extra_covariates:
   - "Number of vehicles in household"
 ```
 
-The `estimands` section describes the types of estimands we are interested in, here Average Interaction Effects (`AIE`). We would like to scan both orders 2 and 3.
+The `estimands` section describes the types of estimands we are interested in, here interactions are defined by the Average Interaction Effects (`AIE`). We would like to scan both orders 2 and 3.
 
-The `variants` section contains groups of variants for which the interactions will be generated independently. Here there is only one group called `TF`, and it contains two subgroups: `variants_1` and `variants_2`. This means that combinations of `variants_1` and `variants_2` will define the interactions under investigation. Here they are simply (1:238411180:T:C, 2:14983:G:A) and (3:3502414:T:C, 2:14983:G:A). Why did we set the interaction `orders` to be both 2 and 3 since only pairwise interactions can be generated here?
+The `variants` section contains groups of variants for which the interactions will be generated independently. Here there is only one group called `TF`, and it contains two subgroups: `variants_1` and `variants_2`. This means that combinations of `variants_1` and `variants_2` will define the interactions under investigation. That is;  (1:238411180:T:C, 2:14983:G:A) and (3:3502414:T:C, 2:14983:G:A). Why did we set the interaction `orders` to be both 2 and 3 since only pairwise interactions can be generated here?
 
[…]

[…]
