### 2026-01-13 cbe53c2
update examples docs and container tags
