+++ .github/workflows/CI.yml
@@ -27,7 +27,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: julia-actions/setup-julia@v2
         with:
-          version: 1
+          version: 1.11
       - uses: actions/cache@v4
         env:
           cache-name: cache-artifacts
