### 2026-01-13 74864c9
port some PR222 changes
