### 2026-01-13 743f20b
port some PR222 changes
