### 2026-01-13 d54028f
more host cleaning and docs formatting
