### 2026-01-13 0e056ae
more host cleaning and docs formatting
