+++ .github/workflows/CI.yml
@@ -26,7 +26,13 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - name: Delete huge unnecessary tools folder
-        run: sudo rm -rf /opt/hostedtoolcache && sudo rm -rf /usr/local/share/boost && sudo rm -rf "$AGENT_TOOLSDIRECTORY"
+        run: |
+          sudo rm -rf /opt/hostedtoolcache
+          sudo rm -rf /usr/local/share/boost
+          sudo rm -rf "$AGENT_TOOLSDIRECTORY"
+          sudo rm -rf /usr/local/lib/android
+          sudo rm -rf /usr/share/dotnet
+          docker system prune -af
       - uses: julia-actions/setup-julia@v2
         with:
           version: 1.11
