### 2026-01-15 a252a42
minor spelling and grammar corrections
