### 2026-01-16 8e71f8a
clean up unused features (e.g. mapping files); match to JOSS branch
