+++ .github/workflows/CI.yml
@@ -48,29 +48,6 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
       - uses: nf-core/setup-nextflow@v1.2.0
-      - name: Free up disk space for simulation tests
-        if: contains(matrix.testrun, 'simulation')
-        run: |
-          echo "Freeing up disk space for ${{ matrix.testrun }}"
-          df -h
-          
-          # Remove unnecessary packages and clean package cache
-          sudo apt-get clean
-          sudo apt-get autoremove -y
-          
-          # Remove large directories that are not needed
-          sudo rm -rf /opt/hostedtoolcache/CodeQL || true
-          sudo rm -rf /opt/hostedtoolcache/Java_Temurin-Hotspot_jdk || true
-          sudo rm -rf /opt/hostedtoolcache/Ruby || true
-          sudo rm -rf /opt/hostedtoolcache/go || true
-          sudo rm -rf /usr/local/lib/android || true
-          sudo rm -rf /usr/share/dotnet || true
-          
-          # Clean Docker system
-          docker system prune -af || true
-          
-          echo "Available disk space after cleanup:"
-          df -h
       - name: Instantiate test environment
         run: julia --project=test --startup-file=no -e'using Pkg;Pkg.instantiate()'
       - name: Run tests
