+++ test/assets/gwas_config.yaml
@@ -1,8 +1,5 @@
 type: gwas
 
-estimands: 
-  - type: ATE
-  
 outcome_extra_covariates:
   - "Number of vehicles in household"
   - "Cheese intake"
