+++ test/Project.toml
@@ -3,7 +3,6 @@ Arrow = "69666777-d1a9-59fb-9406-91d4454c9d45"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-OpenSSL_jll = "458c3c95-2e84-50aa-8efc-19380b2a3a95"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TMLE = "8afdd2fb-6e73-43df-8b62-b1650cd9c8cf"
 TMLECLI = "2573d147-4098-46ba-9db2-8608d210ccac"
