### 2026-01-16 ab1b295
add tests for dense and subset GWAS
