### 2026-01-16 c577021
add tests for dense and subset GWAS
