+++ .github/workflows/CI.yml
@@ -16,6 +16,8 @@ jobs:
       matrix:
         testrun:
           - "ukb_gwas.jl"
+          - "ukb_dense_gwas.jl"
+          - "ukb_subset_gwas.jl"
           - "ukb_estimands_file.jl"
           - "custom_cohort_flat_config.jl"
           - "ukb_interactions_group_config.jl"
