+++ docs/src/targene/tmle.md
@@ -20,3 +20,11 @@ A default TarGene run will thus result in two estimates for each estimand:
 Note however that this is not compulsory. A single estimator can be used and the machine-learning models customised. This is done via the `ESTIMATORS_CONFIG` which is either a simple configuration string or a more elaborate Julia file. Both options are further discussed [here](https://targene.github.io/TMLECLI.jl/stable/tmle_estimation/#Specifying-Estimators).
 
 For example, using a configuration string, a computationally cheaper run with `ESTIMATORS_CONFIG=ose--glmnet` uses only One-Step Estimation with a [GLMNet](https://www.jstatsoft.org/article/view/v033i01) model for both ``Q_Y`` and ``G``.
+
+## Addressing Sampling Bias
+
+In population genetics studies, it is possible that the observed population does not adequately represent the population of interest with respect to the frequency of binary traits like disease. This sampling bias persists in case-control studies in which individuals with a particular phenotype of interest, or cases, are targeted for recruitment to increase the power of the study.
+
+Currently, TarGene only uses estimands that are not invariant to sampling strategy and thus interpretation of such effects are relative to the observed population used for analysis. To obtain estimates that are representative of the true population from which the observations were sampled, one can provide TarGene with the prevalence of the binary phenotype in the true population. With this prevalence parameter, we can perform a case-control weighted estimation procedure to appropriately weight the empirical distribution to match the population of interest thereby minimizing sampling bias brought on by the study design.
+
+This procedure involves the weighting of the learning functions ``Q_Y`` and ``G`` as well as the semi-parametric update step in (w)TMLE. Furthermore, there is no support for the case-control weighting of the OSE and if a custom machine learning algorithm is used for the estimation of ``Q_Y`` or ``G`` it must be able to support weights [MLJ supports weights](https://juliaai.github.io/MLJ.jl/stable/weights/).
