### 2026-01-16 3ac66dc
add gwis docs; remove positivity contraint from new test configs
