+++ docs/src/examples/gwas.md
@@ -57,3 +57,78 @@ The workflow should start running and download the first container image. Once s
 Remember that in TarGene we estimate the effect of allelic changes and we do not assume that the effect of adding one minor allele is the same as the effect of adding the second minor allele. Hence each variant as up to two distinct effect sizes each with their own p-value. The overall p-value (under the `WTMLE_GLM_GLM` section) comes from a multidimensional test asking whether **any** of the allelic changes is non zero. In some cases, only one of the two changes is estimated. This is because there wasn't sufficient data to be confident about assessing the second allelic change (see the `POSITIVITY_CONSTRAINT` in the workflow parameters).
 
 For more on TarGene's outputs, see [Understanding TarGene's Outputs](@ref).
+
+# GWIS
+
+Provided there is sufficient power, one may concern themselves with interaction effects at genome-wide scale or a genome-wide interaction study (GWIS). Whether this query concerns how variants interact with other variants (epistasis) or gene-by-environment interaction, both can be addressed using the GWAS configuration. All that is required for this that differs from the traditional GWAS configuation is some minor additions to the `ESTIMANDS_CONFIG` yaml file.
+
+We can begin by expanding upon our intial estimands configuration. In the first example, we did not need to specify the `estimands` field as they are by default set to be the single order Average Treatment Effects (ATE). Now, that we want more complex estimands we must define the field `estimands` with the subfields `type` and `orders`. Here, we specify that we want to estimate the Average Interaction Effect (AIE) of the second order between each variant and some treatment variable on outcome. These treatment variables are defined under the `extra_treatments` field where we have specified both the genetic variant `rs3693265` and the environmental variable `Hot drink temperature`. This configuration will then define the AIE between each genetic variant the levels of `rs3693265` on outcome as well as the AIE of each genetic variant and the levels of `Hot drink temperature`.
+
+```yaml
+type: gwas
+
+estimands:
+  - type: AIE
+    orders: [2]
+
+extra_treatments:
+  - Hot drink temperature
+  - rs3693265
+
+outcome_extra_covariates:
+  - "Number of vehicles in household"
+  - "Cheese intake"
+```
+
+Futhermore, if we were to add and additional order, say `3`:
+
+```yaml
+type: gwas
+
+estimands:
+  - type: AIE
+    orders: [2,3]
+
+extra_treatments:
+  - Hot drink temperature
+  - rs3693265
+
+outcome_extra_covariates:
+  - "Number of vehicles in household"
+  - "Cheese intake"
+```
+
+We would then have the previously defined estimands as well as a three-point interaction between each variant and the levels of both of the values in `extra_treatments`.
+
+The GWAS can then be run as before:
+
+```bash
+nextflow run https://github.com/TARGENE/targene-pipeline -r v0.13.0 -profile local -c my_gwas.config -resume
+```
+
+And the output would be similar, however the `EFFECT_TYPE` and `TREATMENTS` will, of course, be different. Here is an example of output for the two point interaction between two variants:
+
+```yaml
+- EFFECT_TYPE: "AIE"
+  OUTCOME: Body mass index (BMI)
+  TREATMENTS:
+    - 1:235339932:G:A: "0 => 1"
+      rs3693265: "0 => 1"
+    - 1:235339932:G:A: "1 => 2"
+      rs3693265: "0 => 1"
+    - 1:235339932:G:A: "0 => 1"
+      rs3693265: "1 => 2"
+    - 1:235339932:G:A: "1 => 2"
+      rs3693265: "1 => 2"
+  WTMLE_GLM_GLM:
[…]
