### 2026-01-28 #9603 @ZeyuanSong
Thank you for the updates @olivierlabayle. I’m trying to replicate the GWAS example, but I’m encountering an error at the final step. I’ve attached the error log for reference.

[MyJob.29732149.txt](https://github.com/user-attachments/files/24900000/MyJob.29732149.txt)
