### 2026-01-28 #9603 @olivierlabayle
Thanks @ZeyuanSong , could you inform on which version of the pipeline you are using? For this review it might be useful to use the [joss_review](https://github.com/TARGENE/targene-pipeline/tree/joss_review) branch since this is what will be released next. I have recently successfully ran the examples as described in the documentation on that branch.
Also, unfortunately, the Julia error message is not very informative in that case. Could you provide the content of .command.err and .command.log in your workdir: /cluster/home/zsong07/JOSS-TarGene/work/a6/625a85b8af16f48d4c9520024854a4?
