### 2026-01-30 04f77a6
allow multithreaded estimation inputs; use core feature image; remove tests for new modes
