### 2026-01-30 8b8a399
allow multithreaded estimation inputs; use core feature image; remove tests for new modes
