### 2026-03-03 1a41d67
up documentation and tests for dense and subset GWAS
