### 2026-03-03 bf820f2
up documentation and tests for dense and subset GWAS
