### 2026-03-03 47f7056
remove monomorphic variants from variant testsets
