+++ paper/paper.bib
@@ -75,6 +75,19 @@
   publisher={Springer}
 }
 
+@article{Labayle_TMLE_jl_Targeted_Minimum_2025,
+author = {Labayle, Olivier and Ponting, Chris P. and van der Laan, Mark J. and Khamseh, Ava and Beentjes, Sjoerd Viktor},
+doi = {10.21105/joss.08446},
+journal = {Journal of Open Source Software},
+month = aug,
+number = {112},
+pages = {8446},
+title = {{TMLE.jl: Targeted Minimum Loss-Based Estimation in Julia}},
+url = {https://joss.theoj.org/papers/10.21105/joss.08446},
+volume = {10},
+year = {2025}
+}
+
 @article{purcell2007plink,
   title={PLINK: a tool set for whole-genome association and population-based linkage analyses},
   doi={10.1086/519795},
