### 2026-03-10 #9603 @olivierlabayle
@abhishektiwari I have made the checks, here are the relevant information

- Github Release: v0.13.2 - https://github.com/TARGENE/targene-pipeline/releases/tag/v0.13.2
- Zenodo Archive DOI: https://doi.org/10.5281/zenodo.18935888

Please let me know if I need to do anything else.
