### 2026-03-19 #9603 @abhishektiwari
@olivierlabayle Please see initial edits

**Main Body (Edits in paper.md):**
Line 43: Missing article: "TarGene is Nextflow pipeline" should be "TarGene is a Nextflow pipeline"
Line 25: Incorrect citation strings. It should be `[@purcell2007plink; @yang2011gcta; @loh2018mixed; @zhou2018efficiently]` [see here](https://pandoc.org/MANUAL.html#extension-citations).
Line 38–39: Should be "a framework combining methods from causal inference, machine learning, and semi-parametric statistical theory" (Oxford comma)

**References (Edits in paper.bib):** Please use {} explicitly to capitalise as by default JOSS formats reference titles as sentence case.

Bycroft et al. (2018): "UK biobank" should be "UK Biobank" (proper name of the database)
Canela-Xandri et al. (2018):  "UK biobank." should be "UK Biobank"
Chen & Guestrin (2016): "Xgboost" should be "XGBoost"
Us Research Program Investigators, A. of. (2019): Title to reflect the proper name of the program "All of Us". DOI is incorrect it should be `10.1056/NEJMsr1809937` (case sensitive).

Once done please use to `@editorialbot generate pdf` to check if references are rendered correctly in PDF.
