+++ paper/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @article{canela2018atlas,
-  title={An atlas of genetic associations in UK Biobank},
+  title={An atlas of genetic associations in {UK Biobank}},
   doi={10.1038/s41588-018-0248-z},
   author={Canela-Xandri, Oriol and Rawlik, Konrad and Tenesa, Albert},
   journal={Nature genetics},
@@ -59,7 +59,7 @@
 }
 
 @inproceedings{chen2016xgboost,
-  title={Xgboost: A scalable tree boosting system},
+  title={{XGBoost}: A scalable tree boosting system},
   doi={10.1145/2939672.2939785},
   author={Chen, Tianqi and Guestrin, Carlos},
   booktitle={Proceedings of the 22nd acm sigkdd international conference on knowledge discovery and data mining},
@@ -146,7 +146,7 @@ year = {2025}
 }
 
 @article{bycroft2018uk,
-  title={The UK Biobank resource with deep phenotyping and genomic data},
+  title={The {UK Biobank} resource with deep phenotyping and genomic data},
   doi={10.1038/s41586-018-0579-z},
   author={Bycroft, Clare and Freeman, Colin and Petkova, Desislava and Band, Gavin and Elliott, Lloyd T and Sharp, Kevin and Motyer, Allan and Vukcevic, Damjan and Delaneau, Olivier and O’Connell, Jared and others},
   journal={Nature},
@@ -158,9 +158,9 @@ year = {2025}
 }
 
 @article{all2019all,
-  title={The “All of Us” research program},
+  title={The “{All of Us}” research program},
   doi={10.1056/nejmsr1809937},
-  author={All of Us Research Program Investigators},
+  author={{All of Us Research Program Investigators}},
   journal={New England Journal of Medicine},
   volume={381},
   number={7},
