+++ paper/paper.bib
@@ -159,7 +159,7 @@ year = {2025}
 
 @article{all2019all,
   title={The “{All of Us}” research program},
-  doi={10.1056/nejmsr1809937},
+  doi={10.1056/NEJMsr1809937},
   author={{All of Us Research Program Investigators}},
   journal={New England Journal of Medicine},
   volume={381},
