+++ paper/paper.bib
@@ -68,7 +68,7 @@
 }
 
 @book{van2018targeted,
-  title={Targeted learning in data science},
+  title={{Targeted Learning} in {Data Science}},
   doi={10.1007/978-3-319-65304-4},
   author={Van der Laan, Mark J and Rose, Sherri},
   year={2018},
@@ -137,7 +137,7 @@ year = {2025}
 }
 
 @book{van2011targeted,
-  title={Targeted learning: causal inference for observational and experimental data},
+  title={{Targeted Learning}: {Causal Inference for Observational and Experimental Data}},
   doi={10.1007/978-1-4419-9782-1},
   author={Van der Laan, Mark J and Rose, Sherri and others},
   volume={4},
