### 2026-03-20 #9603 @olivierlabayle
@abhishektiwari I've addressed your comments and released a new version:

- Zenodo archive: https://doi.org/10.5281/zenodo.19127728
- Github release: https://github.com/TARGENE/targene-pipeline/releases/tag/v0.13.3
