### 2026-03-22 #9603 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
