### 2026-03-23 #9603 @Kevin-Mattheus-Moerman
@olivierlabayle As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which require your attention:

👉 Your project appears to lack contributing guidelines at the moment (or at least I do not see them from the main repository page or readme). Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

👉 Please spell out `USA` as `United States of America`
- [x] Copyedit pass/checked for typos
