### 2026-03-24 56cd634
Add contributing section to README
Added contribution guidelines section with a link.
