### 2026-03-24 #9603 @olivierlabayle
Hi @Kevin-Mattheus-Moerman ,

- There is a contributing guide available from the docs: [contributing](https://targene.github.io/targene-pipeline/stable/developer_guide/contribution_guide/). I have now added a link to the relevant page in the [README.md](https://github.com/TARGENE/targene-pipeline?tab=readme-ov-file)
- USA -> United States of America: done

New github release: [v0.13.4](https://github.com/TARGENE/targene-pipeline/releases/tag/v0.13.4)
New Zenodo Archive DOI : [https://doi.org/10.5281/zenodo.19202203](https://doi.org/10.5281/zenodo.19202203)
